% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-types.R
\name{basic-mission}
\alias{basic-mission}
\alias{get_basic_mission}
\alias{is_bomber}
\alias{is_fighter}
\alias{is_ground_attack}
\alias{is_multi_mission}
\alias{is_tanker}
\alias{is_trainer}
\alias{is_transport}
\title{Basic Mission Symbol}
\usage{
is_ground_attack(x, ...)

is_bomber(x, ...)

is_transport(x, ...)

is_fighter(x, ...)

is_tanker(x, ...)

is_multi_mission(x, ...)

is_trainer(x, ...)

get_basic_mission(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{...}{Additional optional arguments to be passed onto \code{grepl}.}
}
\description{
Extract or test for a specific basic mission symbol.
}
\details{
The \code{is_Y} functions return TRUE if \code{x} includes the basic mission
symbol \code{Y}, and FALSE otherwise. The function \code{get_basic_mission}
will extract and return the basic mission symbol, if any.
}

