% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_df}
\alias{char_to_df}
\title{Character Vector to Header of Empty Data Frame}
\usage{
char_to_df(x)
}
\arguments{
\item{x}{Input character vector.}
}
\value{
Empty data frame where columns are named \code{x}.
}
\description{
Converts a character vector into the column names of an empty data frame.
Used in conjunction with \code{check_names} parameter of
\code{input_param_checker()}
}
\keyword{internal}
