% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_chkbrownvars.R
\name{chk_bene_id_18900}
\alias{chk_bene_id_18900}
\title{chk_bene_id_18900: Test Beneficiary ID}
\usage{
chk_bene_id_18900(x)
}
\description{
Called when testing a MDS data.frame is Brown U. canonical
}
