% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_date}
\alias{convert_date}
\title{Convert to Acceptable Date}
\usage{
convert_date(x, convert_to = "month")
}
\arguments{
\item{x}{Input vector of class \code{Date}.}

\item{convert_to}{Default: \code{"month"}. String value indicating the
conversion. Current acceptable conversions are \code{"day"}, \code{"month"},
\code{"quarter"}, \code{"semiannual"}, \code{"annual"}.}
}
\value{
Converted \code{Date} vector of class \code{mdpms.Date} with
attributes \code{duration} (Lubridate concept), \code{current_conversion},
and \code{possible_conversions}.
}
\description{
Converts a \code{Date} vector into into its equivalent daily, monthly, yearly,
etc... \code{Date} vector.
}
