% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{correlation}
\usage{
correlation(Year)
}
\arguments{
\item{year}{When do you want to know the correlation matrix of numeric data?}
}
\description{
This function returns a correlation matrix of numeric value of a year.
}
\examples{
correlation()
}
\keyword{correlation}
