% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_cps.R
\name{cpt_cps}
\alias{cpt_cps}
\title{Compute Cognitive-Performance Scale Score}
\usage{
cpt_cps(x, .report=T)
}
\arguments{
\item{x}{A data.frame class object, with M3G0110 items.}

\item{report}{Generates a short report on computed ADL score}
}
\value{
A dataframe with added column adl_score
}
\description{
Scans MDS dataframe for 'std' named BIMS / CPS items and
computes cps for each assessment.
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
cpt_adl28(mds_dta)

}
\references{
ADD LATER
}
