% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_stay.R
\name{cpt_stay}
\alias{cpt_stay}
\title{cpt_stay: Flag residents by stay length}
\usage{
cpt_stay(mds_obj,
                .stay_min = 100L,
                .stay_maxout,
                .report=T)
}
\arguments{
\item{.stay_min}{A integer of length one. The length of stay to flag.
100 days is the default.}

\item{.stay_maxout}{A integer of length one. The maximum allowable days
outside of nursing home. 10 days is default.}

\item{.report}{A logical indicating a short report on cohort construction.
Default is true.}

\item{mds_obj}{A data.frame class object which is "mds canon".
See ?mds_as_canon}

\item{.keep_all}{A logical indicates whether to keep all the patients, or
if false only those meeting .stay_* rules. Default is false.}
}
\value{
A dataframe with one row per resident, cohort entry, cohort exit,
flag for rules equals T if .stay_* rules met,
}
\description{
Scans a data.frame of minimum dataset assessments, assigns an
indicator according to an arbitrary length of stay. 100 days with no more
than 10 days out is a standard "long-stay" definition.
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
longstay <- cpt_stay(mds_dta, .keep_all=T)

}
\references{
ADD LATER
}
