% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_analyses.R
\name{define_analyses}
\alias{define_analyses}
\title{Set MD-PMS Analyses Definitions}
\usage{
define_analyses(deviceevents, exposure = NULL, algorithms = NULL,
  date_level = "month", device_levels = "device_1", event_levels = NULL,
  times_to_calc = NULL, prior = NULL)
}
\arguments{
\item{key}{Character name of (uniquely identifying) primary key variable in
\code{data_frame}. Class must be character or numeric.

Example: \code{"key_ID"}

Default: \code{NULL} will create a key variable.}
}
\value{
A list of defined analyses of class \code{mdpms.analyses}.
Attributes are as follows:
\describe{
  \item{key}{Original variable name for \code{key}}
  \item{time}{Original variable name for \code{time}}
  \item{device_hierarchy}{Vector of original variable names for
  \code{device_hierarchy} with converted variable names correspondingly
  named.}
  \item{event_hierarchy}{Vector of original variable names for
  \code{event_hierarchy} with converted variable names correspondingly
  named.}
  \item{covariates}{Vector of original variable names for
  \code{covariates} with converted variable names correspondingly
  named.}
}
}
\description{
Define and enumerate analyses based on MS-PMS device-event data frames and
(optionally) MD-PMS exposure data frames.
}
\examples{
## Need to write!

}
