% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_analyses.R
\name{define_analyses_dataframe}
\alias{define_analyses_dataframe}
\title{Create Data Frame from Analyses Definitions}
\usage{
define_analyses_dataframe(inlist)
}
\arguments{
\item{inlist}{Object of class \code{mds_das}}
}
\value{
A data frame with each row repesenting an analysis.
}
\description{
Returns a data frame summarizing all defined analyses from the
\code{mds_das} object.
}
