% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_device_events.R
\name{device_event_class}
\alias{device_event_class}
\title{Title}
\usage{
device_event_class(data_frame, key, time, device_hierarchy, event_hierarchy)
}
\arguments{
\item{data_frame}{The input described

Long description
blah blah blah}
}
\value{
The object of something.
}
\description{
Short description
}
\examples{
helloworld()

}
