% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fNA}
\alias{fNA}
\title{Min/Max With All NA's Allowed
Min and Max functions that allows the vector \code{x} to be \code{NA}, where
\code{f} is \code{min} or \code{max}}
\usage{
fNA(x, f)
}
\arguments{
\item{x}{Numeric input vector}

\item{f}{Either \code{min} or \code{max}}
}
\value{
Numeric value corresponding to \code{min} or \code{max}, or \code{NA}
}
\description{
Min/Max With All NA's Allowed
Min and Max functions that allows the vector \code{x} to be \code{NA}, where
\code{f} is \code{min} or \code{max}
}
\keyword{internal}
