% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{get_mds_components}
\alias{get_mds_components}
\title{Extract MDS Components}
\usage{
get_mds_components(x)
}
\arguments{
\item{x}{A Character string.}
}
\value{
A character vector with the following six components:
  \describe{
    \item{\code{"status_prefix"}}{status prefix;}
    \item{\code{"modified_mission"}}{modified mission;}
    \item{\code{"basic_mission"}}{basic mission;}
    \item{\code{"vehicle_type"}}{vehicle type;}
    \item{\code{"design_number"}}{design number;}
    \item{\code{"series_letter"}}{series letter.}
  }
}
\description{
Extract the six basic components from a given MDS.
}
\examples{
# A single MDS
get_mds_components("F-017")

# Multiple MDS
mds <- c("YEH-060B", "F17", "F-17", "F017", "F-017", "YRAH-066A")
plyr::ldply(mds, get_mds_components)
}

