% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifelse_cp}
\alias{ifelse_cp}
\title{Class-Preserving If-Else
If-Else function preserving \code{T} & \code{F} classes}
\usage{
ifelse_cp(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode}

\item{yes}{return values for true elements of \code{test}.}

\item{no}{return values for false elements of \code{test}.}
}
\value{
A vector of the same length and attributes as \code{test}.
}
\description{
Class-Preserving If-Else
If-Else function preserving \code{T} & \code{F} classes
}
\keyword{internal}
