% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{is_valid_mds}
\alias{add_hyphen}
\alias{is_invalid_mds}
\alias{is_valid_mds}
\alias{standardize_mds}
\alias{strip_hyphen}
\alias{strip_leading_zeros}
\title{Mission Design Series (MDS)}
\usage{
is_valid_mds(x, ...)

is_invalid_mds(x, ...)

add_hyphen(x, ...)

strip_hyphen(x, ...)

strip_leading_zeros(x, ...)

standardize_mds(x, hyphen = TRUE, ...)
}
\arguments{
\item{x}{A character vector.}

\item{...}{Additional optional arguments.}

\item{hyphen}{Logicial indicating whether or not to include a hyphen before
the design number (if one is not already present). Only used in
\code{standardize_mds}.}
}
\description{
Various functions for working with MDS codes.
}
\examples{
mds <- c("YEH-60B", "YEH-60B1", "AYEH-60B", "F-16A", "F16", "F016", "YRAH-66A")

is_valid_mds(mds)

mds \%>\%
  strip_leading_zeros \%>\%
  strip_hyphen

mds \%>\%
  strip_leading_zeros \%>\%
  add_hyphen
}

