% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_canon.R
\name{mds_as_canon}
\alias{mds_as_canon}
\title{mds_as_canon: Test object for canon MDS file}
\usage{
mds_as_canon(mds_obj, .report=T)
}
\arguments{
\item{mds_obj}{A data.frame class object.}

\item{.report}{A logical vector, indicated whether to output to console}
}
\value{
Prints output to console, if canon returns data.frame with attribute
}
\description{
Called when testing an object is MDS canonical
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
mds_as_canon(mds_dta)

}
