% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_canon.R
\name{mds_canon}
\alias{mds_canon}
\title{mds_canon: Test object for canon MDS file}
\usage{
mds_canon(x, .report=T)
}
\arguments{
\item{x}{A data.frame class object.}
}
\value{
Prints output to console
}
\description{
Called when testing an object is MDS canonical
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
mds_canon(mds_dta)

}
