% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_chkfile.R
\name{mds_chkfile}
\alias{mds_chkfile}
\title{mds_chkfile: Test object for canon MDS file}
\usage{
mds_chkfile(x, .report=T)
}
\arguments{
\item{x}{A data.frame class object.}
}
\value{
Prints output to console
}
\description{
Called when testing an object is MDS canonical
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
mds_chkfile(mds_dta)

}
