% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_chkmds.R
\name{mds_chkmds}
\alias{mds_chkmds}
\title{mds_chkmds: Generate a report on MDS data.frame}
\usage{
mds_chkmds(mds_obj, .report=T)
}
\arguments{
\item{mds_obj}{A data.frame class object.}
}
\value{
Prints output to console
}
\description{
Called when testing an object is MDS canonical
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
mds_chkmds(mds_dta)

}
