% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_fmt.R
\name{mds_fmt}
\alias{mds_fmt}
\title{mds_fmt: Format an MDS file}
\usage{
mds_fmt(mds_obj, .quietly=F)
}
\arguments{
\item{mds_obj}{A data.frame class object with MDS items.}

\item{.quietly}{If FALSE, will print some output to console}
}
\value{
A tibble class dataframe
}
\description{
Runs a series of functions to reformat and edit a file to make
final analytical file
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_dta
mds_fmt(mds_dta)

}
