% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_rpt_xls.R
\name{mds_rpt_xls}
\alias{mds_rpt_xls}
\title{Descriptive Report of MDS File}
\usage{
mds_rpt_xls(mds_obj, 'mds.xlsx', .quietly=F)
}
\arguments{
\item{mds_obj}{A data.frame class object with MDS items.}

\item{filepath}{Path to save excel file}

\item{.quietly}{If FALSE, will print some output to console}
}
\value{
No return
}
\description{
Outputs select analyses of an MDS data.frame object
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
mds_rpt_xls(mds_dta, 'mds.xslx')

}
