% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_std.R
\docType{data}
\name{mds_std}
\alias{mds_std}
\title{mds_std}
\format{A data frame with 346342 rows and 8 variables:
\describe{
  \item{var}{Variable Name}
  \item{label}{Variable Label}
  \item{desc}{Variable Description}
  \item{type}{Variable type}
  \item{length}{Variable length}
}}
\source{
\url{https://data.cdc.gov/dataset/Deaths-in-122-U-S-cities-1962-2016-122-Cities-Mort/mr8w-325u}
}
\usage{
mds_std
}
\description{
Standard MDS data dictionary; includes list of variables and attributes. If
not specified otherwise, all tests / format checks will use this file for
comparing.
}
\keyword{datasets}
