% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mds_ts}
\alias{mds_ts}
\title{Example List of mds_ts Time Series Objects}
\format{A list of 3 elements each of class \code{mds_ts}}
\source{
See \code{?maude} and \code{?sales} for source device-event and
exposure data. See \code{?mds::time_series} for how to generate \code{mds_ts}
time series.
}
\usage{
mds_ts
}
\description{
An example list of time series objects (class \code{mds_ts}) generated using
the \code{mds} package.
}
\keyword{datasets}
