% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm.R
\name{mmm}
\alias{mmm}
\title{On the testing of multiple hypothesis in sliced inverse regression}
\usage{
mmm(y, x, q = 0.1, nslices = 8)
}
\arguments{
\item{y}{response variable.}

\item{x}{input design matrix of dimension n x p; each row is an observation vector.}

\item{q}{target FDR level. The default value is 0.1.}

\item{nslices}{number of cores used for parallel computation of the mmm_1 to mmm_p. We recommend to use multiple cores which will greatly reduce the computational time.}
}
\value{
mmm_selected: the index of selected variable.

mmm_statistics: the value for mirror statistics.
}
\description{
On the testing of multiple hypothesis in sliced inverse regression
}
\examples{
\donttest{n <- 300  # Number of observations
p <- 1000  # Number of predictors included in model
q <- 60  # Number of true predictors
amplitude = 5 # signal amplitude (for noise level = 1) 


# Generate the variables from a multivariate normal distribution
times <- 1:p
simmma <- 1
rho = 0.2
H <- abs(outer(times, times, "-"))
V <- simmma * rho^H
x = mvrnorm(n, mu = rep(0,p), Simmma=V)

# Generate the response.

a1_nonnull = rep(amplitude/sqrt(n),q)
a2_nonnull = rep(amplitude/sqrt(n),q)
true_index= sample(p,q)
true_index=1:q
a1  = numeric(p)
a2  = numeric(p)
a1[true_index]=a1_nonnull
a2[true_index]=a2_nonnull
y = sin(x \%*\% amat[,1]) + (x\%*\% amat[,2])^3 + rnorm(n, sd = 1);

# Fit the MMM model
fit = mmm(y, x)
}


}
\references{
\url{}
}
\author{
Xin Xing, \email{xin_xing@fas.harvard.edu} Zhigen Zhao, \email{zhaozhg@temple.edu}
}
\keyword{fdr}
\keyword{inverse}
\keyword{regression,}
\keyword{selection,}
\keyword{slice}
\keyword{variable}
