% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_dev}
\alias{next_dev}
\title{Return next level up device
Returns the variable name of the next level in the device hierarchy}
\usage{
next_dev(x)
}
\arguments{
\item{x}{String input of device name, such as \code{"device_1"},
\code{"device_2"}, etc.}
}
\value{
String representation of the next device level variable
}
\description{
Return next level up device
Returns the variable name of the next level in the device hierarchy
}
\keyword{internal}
