% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_funcs.R
\name{numeric_fctr}
\alias{numeric_fctr}
\title{numeric_fctr: Compute number (proportion as percent)}
\usage{
numeric_fctr(x, .digits=2, .referent = TRUE)
}
\arguments{
\item{x}{A numeric or factor class vector}

\item{.digits}{Significant figures to round to}

\item{.referent}{Value to compute no. (%) for}
}
\value{
A character string of length 1, 'No. (%)'
}
\description{
Given a single numeric vector or factor,
will compute no. equal to a .referent value and percent of total,
returning results in string. Mostly called by other functions.
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
numeric_fctr(mds_dta$M3A0800, .referent==1)

}
