% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{oldest}
\usage{
oldest(Year)
}
\arguments{
\item{year}{Which year do you want to know the oldest player?Defaults to 2000}
}
\description{
This function gives you the oldest player during a year.
}
\examples{
oldest()
}
\keyword{age}
