% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales}
\alias{sales}
\title{Simulated Device Sales Data}
\format{A data frame with 360 rows and 4 variables:
\describe{
  \item{device_name}{Name of the device mapped from the \code{maude} dataset.}
  \item{region}{Geographical region mapped from the \code{maude} dataset.}
  \item{sales_month}{The month of sales.}
  \item{sales_volume}{The volume of sales.}
}}
\source{
Random normal distribution using \code{rnorm()}.
}
\usage{
sales
}
\description{
A dataset containing simulated monthly sales by device and country for
devices reported in the \code{maude} dataset. For package usage examples,
this data serves as a proxy for exposures. Data were generated using a
random normal distribution weighted by the number of reported events by
device and country.
}
\keyword{datasets}
