% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessfulDC.R
\name{SuccessfulDC}
\alias{SuccessfulDC}
\title{SuccessfulDC: Find those who are discharged to community}
\usage{
SuccessfulDC(x, '2018-01-01', '2018-10-01')
}
\arguments{
\item{x}{A dataframe class object, must contain
"bene_id_18900", "dmdate", "dmatype", "M3A2100"}

\item{.preindex}{Date class object of length 1, start window to identify}

\item{.postindex}{Date class object of length 1, stop window to identify}
}
\value{
an object of class data.frame, limited to those without discharge
to community in window
}
\description{
Queries a standardized MDS cohort dataset to identify those who
were discharged to community
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
start <- lubridate::ymd('2017-01-01')
stop <- lubridate::ymd('2018-01-01')
mds_dta2 <- SuccessfulDC(mds_dta, start, stop)

}
