% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_analyses.R
\name{summary.mds_das}
\alias{summary.mds_das}
\title{Summarize a Collection of MD-PMS Defined Analyses
Prints basic counts and date ranges by various analysis factors as defined in
the original \code{define_analyses()} call.}
\usage{
\method{summary}{mds_das}(object, ...)
}
\arguments{
\item{object}{A MD-PMS Defined Analyses object of class \code{mds_das}}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
List of analyses counts and date ranges.
}
\description{
Summarize a Collection of MD-PMS Defined Analyses
Prints basic counts and date ranges by various analysis factors as defined in
the original \code{define_analyses()} call.
}
