% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series.list}
\alias{time_series.list}
\title{Generate Time Series from Defined Analyses}
\usage{
\method{time_series}{list}(analysis, ...)
}
\arguments{
\item{analysis}{A \code{list} containing defined analyses, each of class
\code{mds_da}.}

\item{...}{Further arguments passed into \code{\link{time_series.mds_da}}.}
}
\value{
A list of time series objects.
}
\description{
Converts a list of defined analyses (class \code{list}) to a list
of time series objects.
}
\seealso{
\code{\link{time_series.mds_da}}
}
