% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series.mds_das}
\alias{time_series.mds_das}
\title{Generate Time Series from Defined Analyses}
\usage{
\method{time_series}{mds_das}(analysis, ...)
}
\arguments{
\item{analysis}{An object of class \code{mds_das}.}

\item{...}{Further arguments passed into \code{\link{time_series.mds_da}}.}
}
\value{
A list of time series objects.
}
\description{
Converts a collection of defined analyses (class \code{mds_das}) to a list
of time series objects.
}
\seealso{
\code{\link{time_series.mds_da}}
}
