% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Generate Time Series from Defined Analyses}
\usage{
time_series(inlist, deviceevents, exposure = NULL)
}
\arguments{
\item{inlist}{Character name of (uniquely identifying) primary key variable in
\code{data_frame}. Class must be character or numeric.

Example: \code{"key_ID"}

Default: \code{NULL} will create a key variable.}
}
\value{
A standardized MD-PMS data frame of class \code{mdpms.deviceevents}.
Rows are deduplicated. Attributes are as follows:
\describe{
  \item{key}{Original variable name for \code{key}}
}
}
\description{
Converts defined analyses into a time series data frame. Accepts both
a collection of analyses (Class \code{mdpms.define_analyses}) or a single
analysis (Class \code{mdpms.define_analysis}).
}
\examples{
## Need to write!

}
