% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toppoints.R
\name{toppoints}
\alias{toppoints}
\title{toppoints}
\usage{
toppoints(Year)
}
\arguments{
\item{year}{Which year do you want to know the player who got the most scores? Defaults to 2000.}
}
\description{
This function gives you the player who got the most points during a year.
}
\examples{
toppoints()
}
\keyword{points}
