% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VEHICLE_TYPE}
\alias{BASIC_MISSION}
\alias{DESIGN_NUMBER}
\alias{MDS_REGEX}
\alias{MODIFIED_MISSION}
\alias{SERIES_LETTER}
\alias{STATUS_PREFIX}
\alias{VEHICLE_TYPE}
\title{Constants}
\format{An object of class \code{character} of length 1.}
\usage{
VEHICLE_TYPE

BASIC_MISSION

MODIFIED_MISSION

DESIGN_NUMBER

SERIES_LETTER

STATUS_PREFIX

MDS_REGEX
}
\description{
Constants defining various regular expressions related to the various
components of a mission design series (MDS).
}
\details{
A basic MDS should contain (at least) six components:
  \enumerate{
    \item{Vehicle type}
    \item{}
    \item{}
    \item{}
    \item{}
    \item{}
  }
The components are actually presented in the order (6)(3)(2)(1)-(4)(5).
}
\keyword{datasets}

