% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{input_param_checker}
\alias{input_param_checker}
\title{Check Input Parameters}
\usage{
input_param_checker(x, check_class = NULL, check_names = NULL,
  null_ok = T)
}
\arguments{
\item{x}{Single object name, character, or vector of characters}

\item{check_class}{Name of correct class in character format}

\item{check_names}{Optional: Data frame object}

\item{null_ok}{Default: \code{T}. Allow \code{x} to be \code{NULL}. If
\code{x} is \code{NULL}, bypass checking.}
}
\value{
Stop error, if an error is found. Else nothing.
}
\description{
Verifies correct class and, optionally, verifies existence in a data frame
}
