% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{input_param_checker}
\alias{input_param_checker}
\title{Check input parameters}
\usage{
input_param_checker(x, check_class = NULL, check_names = NULL)
}
\arguments{
\item{x}{Object name, character, or vector of characters}

\item{check_class}{Name of correct class in character format}

\item{check_names}{Optional: Data frame object}
}
\value{
Stop error or nothing
}
\description{
Verifies correct class and, optionally, verifies existence in a data frame
}
