% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{MDS}
\alias{MDS}
\title{Plot MDS for a DESeq2 object}
\usage{
MDS(rld)
}
\arguments{
\item{rld}{A DESeqTransform object}
}
\description{
Plot MDS for a DESeq2 object
}
\examples{
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
MDS(rld)
}
