% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_std.R
\docType{data}
\name{mds_std}
\alias{mds_std}
\title{mds_std}
\format{A data frame with XX rows and 5 variables:
\describe{
  \item{Var}{Variable Name}
  \item{Label}{Variable Label}
  \item{Type}{Variable type}
  \item{Length}{Variable length}
  \item{Format}{Variable Special Format, e.g. Date}
  \item{Description}{Variable Description}
}}
\usage{
mds_std
}
\description{
Standard MDS data dictionary; includes list of variables and attributes. If
not specified otherwise, all tests / format checks will use this file for
comparison.
}
\keyword{datasets}
