% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Generate Time Series from Defined Analysis or Analyses}
\usage{
time_series(analysis, ...)
}
\arguments{
\item{analysis}{An object of class \code{mds_da} or \code{mds_das}.}

\item{...}{Further arguments passed onto \code{time_series} methods,
primarily \code{time_series.mds_da()}.}
}
\value{
A time series object or list of time series objects.
}
\description{
Generic function for calculating time series. See
\code{\link{time_series.mds_da}} or \code{\link{time_series.mds_das}} as
applicable.
}
\seealso{
\code{\link{time_series.mds_da}}, \code{\link{time_series.mds_das}}
}
